// start-microfrontends/apps/cofepris/src/app/application/shared/models/shared2606/datos-solicitud.model.ts
export interface Catalogo {
  id: number;
  descripcion: string;
  clave?: string;
  tam?: string;
  dpi?: string;
  scianDescription?: string;
}
// ...
export interface TablaScianConfig {
  clave: string;
  descripcion: string;
}
// ...
export interface TablaOpcionConfig {
  fechaCreacion: string;
  mercancia: string;
  cantidad: string;
  proveedor: string;
   /** RFC Sanitario - Registro Federal de Contribuyentes */
   rfcSanitario: string;
  
   /** Denominación o Razón Social */
   denominacionRazon: string;
   
   /** Correo Electrónico */
   correoElectronico: string;
   
   /** Código Postal */
   codigoPostal: string;
   
   /** Estado */
   estado: string;
   
   /** Municipio o Alcaldía */
   municipioAlcaldia: string;
   
   /** Localidad */
   localidad: string;
   
   /** Colonia */
   colonia: string;
   
   /** Calle y Número */
   calleYNumero: string;
   
   /** Calle */
   calle: string;
   
   /** Código de área telefónica (LADA) */
   lada: string;
   
   /** Número de teléfono */
   telefono: string;
   
   /** Aviso */
   aviso: string;
   
   /** Licencia Sanitaria */
   licenciaSanitaria: string;
   
   /** Régimen */
   regimen: string;
   
   /** Aduanas de Entrada */
   adunasDeEntradas: string;
   
   /** Aeropuerto */
   aeropuerto: boolean;
   
   /** Aeropuerto Dos */
   aeropuertoDos: boolean;
   
   /** Público */
   publico: string;
   
   /** RFC del Representante */
   representanteRfc: string;
   
   /** Nombre del Representante */
   representanteNombre: string;
   
   /** Apellido Paterno */
   apellidoPaterno: string;
   
   /** Apellido Materno */
   apellidoMaterno: string;
   
   /** Régimen de la Mercancía */
   regimenLaMercancia: string;
   
   /** Aduana */
   aduana: string;
   
   /** Mercancías */
   mercancias: TablaMercanciasDatos[];

   scian: TablaScianConfig[];
   
   /** Manifiesto */
   manifesto: string;
   
   /** Manifestos Casilla de Verificación */
   manifiestosCasillaDeVerificacion: boolean;

   id_solicitud?: number,
  fecha_creacion?: string,
 }
// ...
export interface TablaMercanciasDatos {
  denominacionEspecificaProducto?: string | undefined;
  clasificacionProducto: string;
  paisDestino?: string;
  paisDeOrigen?: string;
  claveClasificacionProductoObj ?: Catalogo | undefined;
  especificarClasificacionObj ?: Catalogo | undefined;
  tipoProductoObj ?: Catalogo | undefined;
  formaFarmaceuticaObj ?: Catalogo | undefined;
  estadoFisicoObj ?: Catalogo | undefined;
  cantidadUMCObj ?: Catalogo | undefined;
  especificarClasificacionProducto: string;
  denominacionComunInternacional?: string;
  PorcentajeDeConcentracion?: string;
  valorComercial?: string;
  marcaComercialODenominacionDistintiva?: string;
  denominacionDistintiva?: string;
  denominacionComun?: string;
  formaFarmaceutica?: string;
  estadoFisico: string;
  fraccionArancelaria: string;
  descripcionFraccion?: string;
  unidadMedidaComercializacion?: string;
  cantidadUMC: string;
  cantidadUmcValor?:string;
  unidadMedidaTarifa: string;
  cantidadUMT: string;
  cantidadUmtValor?:string;
  presentacion: string;
  numeroRegistroSanitario: string;
  paisOrigen: string;
  paisProcedencia: string;
  tipoProducto: string;
  usoEspecifico: string;
  detallarUsoEspecifico?:string;
  numeroDePiezasAFabricar?:string;
  descripcionNumeroDePiezas?:string;
  numeroCAS?:string;
  cantidadDeLotes?:string;
  kgPorLote?:string;
  paisDeDestino?:string;
  denominacionCumonInternacional?:string;
  caducidad?:string;
  marcaComercialDenominacion?:string;
  especifique?: string;
  especifiqueForma?: string;
  especifiqueEstado?:string;
  especifiqueObligatorio?:string;
  fechaDeMovimiento?: string;
  paisDeOriginDatos?: string[];
  paisDeProcedenciaDatos?: string[];
  id?: number
  modelo?: string;
  descripcionDelProducto?: string;
  usoEspecificoDatosClave?: string[];
  paisProcedenciaDatosClave?: string[];
  paisOriginDatosClave?: string[];
}
// ...
export interface DatosSolicitudFormState {
  rfcSanitario: string;
  denominacionRazon: string;
  correoElectronico: string;
  codigoPostal: string;
  estado: string | number;
  municipioAlcaldia: string;
  localidad: string;
  colonia: string;
  calleYNumero?: string;
  calle: string;
  lada: string;
  telefono: string;
  aviso: string;
  licenciaSanitaria: string;
  regimen: string | number;
  adunasDeEntradas?: string | number;
  aeropuerto: boolean;
  aeropuertoDos?: boolean;
  publico: string;
  representanteRfc: string;
  representanteNombre: string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  marca?: string;
  especifique?: string;
  claveDeLos?: string;
  fechaDeFabricacio?: string;
  fechaDeCaducidad?: string;
  regimenLaMercancia?:string
  aduana?:string
  manifesto?: boolean;
  manifiestosCasillaDeVerificacion?: boolean;
  rfcRepresentante?:string;
}
// ...
export interface MercanciaForm {
  clasificacionProducto: string;
  especificarClasificacionProducto: string;
  denominacionEspecificaProducto: string;
  denominacionDistintiva: string;
  denominacionComun: string;
  tipoProducto: string;
  formaFarmaceutica?: string;
  estadoFisico: string;
  fraccionArancelaria: string;
  descripcionFraccion: string;
  cantidadUmtValor: string;
  cantidadUmt: string;
  cantidadUmcValor: string;
  cantidadUmc: string;
  presentacion: string;
  numeroRegistroSanitario: string;
  fechaDeMovimiento: string;
  paisDeOriginDatos: string[];
  paisDeProcedenciaDatos: string[];
  usoEspecifico?: string[]
  marca?: string;
  especifique?: string;
  especifiqueForma?: string;
  claveDeLos?: string;
  fechaDeFabricacio?: string;
  fechaDeCaducidad?: string;
  especifiqueObligatorio?:string;
  especifiqueEstado?:string;
  modelo?:string;
  descripcionDelProducto?:string;
}